requirements.txt
absl-py==0.11.0
addict==2.4.0
aiohttp==3.8.1
aiosignal==1.2.0
alembic==1.7.5
amass==0.1.0
antlr4-python3-runtime==4.8
anyio==3.3.0
-e git+https://github.com/NVIDIA/apex@5cb187f39bd9bd9a1a7d58b9ddd517684b962cea#egg=apex
appdirs==1.4.4
argon2-cffi==20.1.0
astor==0.8.1
astunparse==1.6.3
async-generator==1.10
async-timeout==4.0.2
attrs==20.3.0
audioread==2.1.9
Babel==2.9.1
backcall==0.2.0
bleach==3.2.1
cachetools==4.1.1
certifi==2019.11.28
cffi==1.14.4
chamfer-distance @ git+https://github.com/otaheri/chamfer_distance@291c8d338c2dca3a6cbe2df1b891f9fdcc8a411b
chardet==3.0.4
charset-normalizer==2.0.12
click==8.0.3
cmake==3.21.1.post1
color-transfer==0.1
commonmark==0.9.1
configer @ git+https://github.com/nghorbani/configer@8cd1e3e556d9697298907800a743e120be57ac36
configparser==5.0.1
cPython==0.0.6
cycler==0.10.0
Cython==0.29.23
dataclasses==0.6
dbus-python==1.2.16
decorator==4.4.2
defusedxml==0.6.0
deprecation==2.1.0
-e git+ssh://git@github.com/zju3dv/disprcnn.git@7a4d1cff815edcc7e7f5032ccd9b306ef71ab76a#egg=disprcnn
distro==1.4.0
docker-pycreds==0.4.0
easydict==1.9
entrypoints==0.3
Flask==2.0.2
flatbuffers==1.12
freetype-py==2.2.0
frozenlist==1.3.0
fsspec==2022.3.0
future==0.18.2
fvcore==0.1.2.post20201122
gast==0.3.3
gitdb==4.0.9
GitPython==3.1.27
glcontext==2.2.0
google-auth==1.23.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
graphviz==0.19.1
greenlet==1.1.2
grpcio==1.32.0
h5py==2.10.0
human-body-prior==0.9.3.0
Hydra==2.5
hydra-core==1.2.0.dev5
idna==2.8
imageio==2.9.0
imageio-ffmpeg==0.4.2
importlib-metadata==4.10.0
importlib-resources==5.4.0
install==1.3.4
iopath==0.1.9
ipykernel==5.3.4
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
itsdangerous==2.0.1
jedi==0.17.2
Jinja2==3.0.3
joblib==0.17.0
json5==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-core==4.7.0
jupyter-packaging==0.10.4
jupyter-server==1.10.2
jupyterlab==3.1.2
jupyterlab-pygments==0.1.2
jupyterlab-server==2.6.2
jupyterlab-widgets==1.0.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
kornia==0.5.11
librosa==0.8.1
llvmlite==0.36.0
Mako==1.1.6
mara-db==4.7.1
mara-page==1.6.0
mara-pipelines==3.1.2
Markdown==3.3.3
MarkupSafe==2.0.1
matplotlib==3.3.3
mesh-to-sdf==0.0.12
mistune==0.8.4
moderngl==5.6.2
more-itertools==8.12.0
moviepy==2.0.0.dev2
mpipe==1.0.8
multidict==6.0.2
multimethod==1.6
munch==2.5.0
natsort==7.1.1
nbclassic==0.3.1
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
networkx==2.5
ninja==1.10.2.2
notebook==6.1.5
numba==0.53.1
numpy==1.19.4
oauthlib==3.1.0
omegaconf==2.2.0
open3d==0.13.0
opencv-python==4.4.0.46
opt-einsum==3.3.0
packaging==20.4
paddlepaddle==2.1.3
pandas==1.2.4
pandocfilters==1.4.3
parso==0.7.1
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.1
plotly==4.13.0
pooch==1.5.1
portalocker==2.0.0
proglog==0.1.9
prometheus-client==0.9.0
promise==2.3
prompt-toolkit==3.0.8
protobuf==3.14.0
psutil==5.8.0
psycopg2-binary==2.9.2
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyDeprecate==0.3.2
pyglet==1.5.11
Pygments==2.9.0
PyGObject==3.36.0
pykitti==0.3.1
PyMCubes==0.1.2
pymongo==3.11.3
PyOpenGL==3.1.5
PyOpenGL-accelerate==3.1.5
pyparsing==2.4.7
pypoisson==0.10
pyrender==0.1.43
pyrsistent==0.17.3
python-dateutil==2.8.1
pythondialog==3.5.3
pytorch-ignite==0.2.0
pytorch-lightning==1.6.2
pytorch3d==0.3.0
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
pyzmq==20.0.0
requests==2.22.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
resampy==0.2.2
retrying==1.3.3
rich==12.2.0
rsa==4.6
scikit-image==0.17.2
scikit-learn==0.23.2
scipy==1.5.4
Send2Trash==1.5.0
sentry-sdk==1.5.10
setproctitle==1.2.3
Shapely==1.7.1
shortuuid==1.0.8
six==1.15.0
sklearn==0.0
smmap==5.0.0
smplx==0.1.26
sniffio==1.2.0
SoundFile==0.10.3.post1
SQLAlchemy==1.4.29
SQLAlchemy-Utils==0.38.1
ssh-import-id==5.10
svox-t==0.0.1
svrl0==0.0.0
tabulate==0.8.7
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorboardX==2.2
tensorflow==2.4.1
tensorflow-estimator==2.4.0
termcolor==1.1.0
terminado==0.9.1
testpath==0.4.4
threadpoolctl==2.1.0
tifffile==2020.11.18
tomlkit==0.7.2
torch==1.10.1+cu111
torch-ema==0.3
torchaudio==0.10.1+rocm4.1
torchgeometry==0.1.2
torchmetrics==0.8.1
torchsearchsorted @ file:///data/new_disk/wangla/tmp/torchsearchsorted
torchvision==0.11.2+cu111
tornado==6.1
tqdm==4.53.0
traitlets==5.0.5
transforms3d==0.3.1
trimesh==3.8.14
typing_extensions==4.2.0
UNKNOWN==0.0.0
urllib3==1.25.8
visual-hull==0.1
wandb==0.12.15
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.1.0
Werkzeug==2.0.2
widgetsnbextension==3.5.1
wrapt==1.12.1
xlrd==1.2.0
yacs==0.1.8
yarl==1.7.2
zipp==3.6.0
